Replace Talon FX interfaces with Talon SRX in sim
This commit changes the Talon FX motor controllers to use
the Talon SRX interfaces in order to support the CTRE sim stuff,
since there is no support for Talon FX sim right now.

In order to do this replacement, we stop the old read threads with the
FX interfaces, modify the shared pointers for the FX interfaces to use
SRX and then restart the thread. Note, this is all done in the sim init
code so it shouldn't affect the real robot.